Encrypt password on provision
